Move params.py to parameter.py
