+++ python/ray/scripts/scripts.py
@@ -15,7 +15,7 @@ from ray.autoscaler.commands import (attach_cluster, exec_cluster,
 import ray.ray_constants as ray_constants
 import ray.utils
 
-from ray.params import RayParams
+from ray.parameter import RayParams
 
 logger = logging.getLogger(__name__)
