+++ test/failure_test.py
@@ -11,7 +11,7 @@ import tempfile
 import threading
 import time
 
-from ray.params import RayParams
+from ray.parameter import RayParams
 import ray.ray_constants as ray_constants
 from ray.utils import _random_string
 import pytest
