+++ test/stress_tests.py
@@ -9,7 +9,7 @@ import pytest
 import time
 
 import ray
-from ray.params import RayParams
+from ray.parameter import RayParams
 import ray.tempfile_services
 import ray.ray_constants as ray_constants
